Add snapshot regression tests
I'm about to make some changes to the alignArrays module. Before I do
that, I wanted to make sure we have a good testing setup in place.

This will make it easier for outside contributors when reporting bugs.
They can add a new test case and either fix or ask for a fix.