Make DRIFT_RANGE dynamic based on height of images
We've added DRIFT_RANGE as a performance boost for the lcs
implementation. Instead of checking infinitely long windows of
similarities, we've been capping the search at 200px.

When images are long however, the 200 range isn't enough. To make large
diffs more correct, we're applying the range dynamically based on the
height of the image. 10% or 200px, whichever is largest.