Replace euclidean distance with YIC NTSC
Our first iteration of measuring the difference of two colors used a
basic euclidean distance. However, this is not very well aligned
with how human vision works, so it produces disappointing results when
applied to real world diffs.

Looking for a better option, I found mapbox's pixelmatch package, which
describes itself as:

> The smallest, simplest and fastest JavaScript pixel-level image
> comparison library, originally created to compare screenshots in
> tests.

This sounds perfect for this application.

The pixelmatch package has functions that take two entire images, and
does a lot more than we actually need here, so instead of pulling it in
as a dependency, I decided to copy and modify the relevant code to
better fit our use-case.

This is a perceived color difference using YIQ NTSC transmission color
space. More information can be found in this paper:

  http://www.progmat.uaem.mx:8080/artVol2Num2/Articulo3Vol2Num2.pdf

pixelmatch has an anti-aliasing detection feature which will exclude
anti-aliased pixels as "different" that I didn't include here, but that
might be worth investigating down the road.

One aspect of this change that I'm not entirely sure of is pixelmatch
has a threshold option that it uses to ignore changes lower than a
specified delta percentage. This threshold option is squared against
the max YIQ difference. Since we are returning a normalized value
between 0 and 1 here, it isn't entirely clear to me if we should be
dividing the difference by the max difference, taking a square root
somewhere, or square the threshold option when using deep compare (in a
different project).

Another aspect that I'm not entirely sure about is that this algorithm
will blend semi-transparent colors with white, which means that white
and fully transparent are reported as the same thing.