Improve performance of similar enough images
I was tracking down a performance issue with a large diff on happo.io
and found that the shortcut we take when images are "similar enough"
wasn't fast enough. The hashing process takes a little while on large
images and we don't really need the hashing for the similarEnough
function to run. I rewrote it so that it would operate on unhashed image
arrays instead, plus made it early-bail so that we won't have to iterate
through the whole image when we already know it's not similar enough.