Improve showing different dimension diffs
If two images were of different size but had overlapping content, we
wouldn't display diffs properly. Instead of highlighting the "added"
areas, we would simply ignore them and show a total/max diff of zero.

It all boiled down to `colorDelta` not handling zero alpha values
properly. I added a shortcut so that if one image has a fully
transparent alpha and the other image doesn't, we show it as fully
diffing. This will work for all web screenshots that we produce from
Happo, but might look a little funky when providing own images through
the api which are semi-transparent. For now I'm happpy to live with that
limitation however, at least this is better than before (even for
semi-transparent images).