Assume background color when adding padding
When two images aren't matching sizes, we add placeholder transparent
pixels. This works in many cases, but when the two images are very
different, the diff tends to put a giant block of magenta either at the
top or the bottom, hiding "real" content underneath. To fix this, I'm
making the placeholder transparent pixels grayish, blended with the
background color, which is assumed to be the top leftmost pixel.

The assumption about the top left pixel could prove too naive, but for
now it's better than the previous approach which included making a
transparent black pixel.