Use rgba 1,1,1,1 as "filler" pixel
Some code we had was assuming that everything fully transparent was
intentionally added by us as filler for missing content on the right of
either the before or after image. To make things more explicit, I'm
using the term "filler pixel" here and setting it to always be 1,1,1,1.
This way the intentions in the code become clearer and at the same time
we have less risk of collisions with a transparent pixel that we do not
control.