Return negative colorDiff if pixel is darker
I am implementing antialiasing detection in our comparison service and I
need to be able to determine if the pixel is darker or lighter. We can
encode this information by using a negative number if the pixel is
darker, and a positive number if the pixel is lighter. We don't need
this information in other places, so we can use Math.abs to normalize.

While I was at it, I switched getDiffPixel over to use
colorDeltaChannels, since colorDelta is deprecated for performance
reasons.