Replace jimp with sharp
This is only used for tests here. We are using sharp in other projects
so I thought it would be nice to be consistent in this repo.

Sharp does not accept URLs like jimp does, so I also decided to remove
the http-server and just read the files off the filessytem in these
tests.